fix:sms delivery failure due to event type title length constraints
